mysite/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
mysqlclient==2.1.1
pipreqs==0.4.11
PyMySQL==1.0.2
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
yarg==0.1.9
